### 2024-11-08 #672 @pachadotdev
redatam
Submitting Author Name: Mauricio Pacha Vargas Sepulveda
Submitting Author Github Handle: @pachadotdev
Repository:  https://github.com/litalbarkai/open-redatam
Sub-directory: rpkg
Submission type: Pre-submission

---

```
Package: redatam
Title: Import 'REDATAM' Files
Version: 2.0.3
    person(
        given = "Mauricio",
        family = "Vargas Sepulveda",
        email = "m.sepulveda@mail.utoronto.ca",
    person(
        given = "Lital",
        family = "Barkai",
        role = "aut"),
    person(
        given = "Arseny",
        family = "Kapoulkine",
        role = "ctb",
        comment = "'pugixml' C++ library"),
    person(
        family = "Republic of Ecuador",
        role = "dtc",
        comment = "Galapagos census data")
    )
    janitor,
    stringi
Depends: R(>= 3.5.0)
Description: Import 'REDATAM' formats into R via the 'Open REDATAM' C++ library
    <https://github.com/litalbarkai/open-redatam> based on De Grande (2016)
    <https://www.jstor.org/stable/24890658>.
License: Apache License (>= 2)
NeedsCompilation: yes
LinkingTo: cpp11
```

REDATAM is a closed-source format for census and survey data. This package is an "archeological" version of the haven package, and allows to read this specific format widely used in Latin America by different govt. statistical offices. With this package, I have been able to convert census data from the 1990s that is not possible with the Redatam software on Windows because of multiple hardware changes in the last 30 years, and this software also reads recent census data (2017-2020) correctly.

Sociologists, Political Scientists and Economists that need census data and an easy way to read it in R (or Python) to fit regression models or different kinds of analysis.

No. There is a "redatamx" that reads a newer format.

-   (If applicable) Does your package comply with our [guidance around _Ethics, Data Privacy and Human Subjects Research_]
-(https://devguide.ropensci.org/policies.html#ethics-data-privacy-and-human-subjects-research)?

Yes.

This package was removed from CRAN for asking about a specific CLANG-ASAN error that took me long to replicate. The error was asked here as well https://stackoverflow.com/questions/79171799/addresssanitizer-error-alloc-dealloc-mismatch-operator-new-vs-free-in-r-packa
