### 2024-11-11 #672 @emilyriederer
Hey @pachadotdev ! This seems like some very cool software and an important goal. Could you please elaborate on how you see this package interacting with the `redatamx` package you mention and the `litalbarkai/open-redatam` package it is forked from? As a general principle, we're unable to consider forks for submission. Would it be possible to contribute this code to the original source repo or restructure it?
