### 2024-11-12 d895519
fix CRAN ASAN-UBSAN warnings
