+++ rpkg/src/redatamlib/readers/FuzzyVariableParser.cpp
@@ -18,8 +18,17 @@ FuzzyVariableParser::FuzzyVariableParser(ByteArrayReader reader,
 void FuzzyVariableParser::ParseAllVariables(vector<Entity> &entities) {
   vector<pair<size_t, size_t>> searchBounds = GetSearchBounds(entities);
 
-  size_t numThreads = std::thread::hardware_concurrency();
-  numThreads = std::min(entities.size(), numThreads);
+  // R-devel suggestion: Default to using all available hardware concurrency
+  size_t maxThreads = std::thread::hardware_concurrency();
+
+  // Then check for _R_CHECK_LIMIT_CORES_ environment variable
+  const char *checkLimitCoresEnv = std::getenv("_R_CHECK_LIMIT_CORES_");
+  if (checkLimitCoresEnv != nullptr &&
+      std::string(checkLimitCoresEnv) == "TRUE") {
+    maxThreads = 2;
+  }
+
+  size_t numThreads = std::min(entities.size(), maxThreads);
 
   size_t chunkSize = entities.size() / numThreads;
 
@@ -41,6 +50,10 @@ void FuzzyVariableParser::ParseAllVariables(vector<Entity> &entities) {
 vector<pair<size_t, size_t>>
 FuzzyVariableParser::GetSearchBounds(vector<Entity> entities) {
   vector<pair<size_t, size_t>> ret;
+
+  if (entities.empty()) {
+    return ret;
+  }
 
   for (size_t i = 0; i < entities.size() - 1; ++i) {
     ret.push_back(
