### 2024-11-13 be017f8
safely set number of cores
