### 2024-11-13 cef6b69
apply R suggestions to pypkg and cpp code
