### 2024-11-13 4f7d3cd
refactor RListExporter
