+++ dev/04-test-equivalence.r
@@ -1,7 +1,7 @@
 library(dplyr)
 library(data.table)
-library(ipumsr
-library(janitor))
+library(ipumsr)
+library(janitor)
 
 # IPUMS data ----
 
@@ -100,6 +100,84 @@ generate_tables <- function(d_control, file_name, country_code, census_year, var
   return(list(d_control_sex = d_control_sex, d_control_age = d_control_age))
 }
 
+generate_tables_r <- function(d_control, country_code, census_year, var_sex, var_age) {
+  # just Chile, it is identical to the other validation function
+  file_name <- "downloads/CP2017CHL/BaseOrg16/CPV2017-16.dic"
+  # Process d_control for sex
+  d_control_sex <- d_control %>%
+    filter(country == country_code, year == census_year) %>%
+    group_by(sex) %>%
+    summarise(n_ipums = sum(perwt)) %>%
+    ungroup() %>%
+    mutate(pct_ipums = n_ipums / sum(n_ipums) * 100)
+
+  # Process d_control for age
+  d_control_age <- d_control %>%
+    filter(country == country_code, year == census_year) %>%
+    group_by(age2) %>%
+    summarise(n_ipums = sum(perwt)) %>%
+    ungroup() %>%
+    mutate(pct_ipums = n_ipums / sum(n_ipums) * 100)
+
+  # Read and clean the external data
+  d_external <- redatam::read_redatam(file_name)
+
+  names(d_external)
+
+  d_sex <- d_external$persona %>%
+    group_by(sex = !!sym(var_sex)) %>%
+    count(name = "n_rdtm") %>%
+    ungroup() %>%
+    mutate(pct_rdtm = n_rdtm / sum(n_rdtm) * 100)
+
+  # Bind columns and calculate differences for sex
+  d_control_sex <- d_control_sex %>%
+    left_join(d_sex) %>%
+    mutate(
+      n_diff = n_ipums - n_rdtm,
+      pct_diff = pct_ipums - pct_rdtm
+    )
+
+  # Bind columns and calculate differences for age
+  d_control_age <- d_control_age %>%
+    left_join(
+      d_external$persona %>%
+        mutate(
+          age2 = case_when(
+            !!sym(var_age) >= 0 & !!sym(var_age) <= 4 ~ 1,
+            !!sym(var_age) >= 5 & !!sym(var_age) <= 9 ~ 2,
+            !!sym(var_age) >= 10 & !!sym(var_age) <= 14 ~ 3,
+            !!sym(var_age) >= 15 & !!sym(var_age) <= 19 ~ 4,
+            !!sym(var_age) >= 20 & !!sym(var_age) <= 24 ~ 12,
+            !!sym(var_age) >= 25 & !!sym(var_age) <= 29 ~ 13,
+            !!sym(var_age) >= 30 & !!sym(var_age) <= 34 ~ 14,
+            !!sym(var_age) >= 35 & !!sym(var_age) <= 39 ~ 15,
+            !!sym(var_age) >= 40 & !!sym(var_age) <= 44 ~ 16,
+            !!sym(var_age) >= 45 & !!sym(var_age) <= 49 ~ 17,
+            !!sym(var_age) >= 50 & !!sym(var_age) <= 54 ~ 18,
+            !!sym(var_age) >= 55 & !!sym(var_age) <= 59 ~ 19,
+            !!sym(var_age) >= 60 & !!sym(var_age) <= 64 ~ 20,
+            !!sym(var_age) >= 65 & !!sym(var_age) <= 69 ~ 21,
+            !!sym(var_age) >= 70 & !!sym(var_age) <= 74 ~ 22,
[…]
