+++ rpkg/DESCRIPTION
@@ -1,7 +1,7 @@
 Package: redatam
 Type: Package
 Title: Import 'REDATAM' Files 
-Version: 2.0.3
+Version: 2.0.4
 Authors@R: c(
     person(
         given = "Mauricio",
