### 2024-11-14 bbb6524
fix CRAN ubsan errors?
