+++ dev/04-test-equivalence.r
@@ -48,7 +48,8 @@ generate_tables <- function(d_control, file_name, country_code, census_year, var
     ungroup() %>%
     mutate(pct_rdtm = n_rdtm / sum(n_rdtm) * 100)
 
-  if (file_name == "downloads/redatam/CP2012BOL/open-redatam-dicx-to-csv/PERSONA.csv.gz") {
+  if (file_name == "downloads/redatam/CP2012BOL/open-redatam-dic-to-csv/PERSONA.csv.gz" |
+      file_name == "downloads/redatam/CP2012BOL/open-redatam-dicx-to-csv/PERSONA.csv.gz") {
     d_sex$sex <- c(2L,1L)
   }
 
@@ -100,9 +101,7 @@ generate_tables <- function(d_control, file_name, country_code, census_year, var
   return(list(d_control_sex = d_control_sex, d_control_age = d_control_age))
 }
 
-generate_tables_r <- function(d_control, country_code, census_year, var_sex, var_age) {
-  # just Chile, it is identical to the other validation function
-  file_name <- "downloads/CP2017CHL/BaseOrg16/CPV2017-16.dic"
+generate_tables_r <- function(d_control, file_name, country_code, census_year, var_sex, var_age) {
   # Process d_control for sex
   d_control_sex <- d_control %>%
     filter(country == country_code, year == census_year) %>%
@@ -124,11 +123,22 @@ generate_tables_r <- function(d_control, country_code, census_year, var_sex, var
 
   names(d_external)
 
+  if (any(file_name %in% c("downloads/redatam/CP2007PER/CP2007PER/BasePub/CPV2007PER_PUB.dic",
+                           "downloads/redatam/CP2007PER/CP2007PER/BasePub/CPV2007PER_PUB.dicx"))) {
+    # rename d_external$poblacio as d_external$persona
+    names(d_external)[which(names(d_external) == "poblacio")] <- "persona"
+  }
+  
   d_sex <- d_external$persona %>%
     group_by(sex = !!sym(var_sex)) %>%
     count(name = "n_rdtm") %>%
     ungroup() %>%
     mutate(pct_rdtm = n_rdtm / sum(n_rdtm) * 100)
+
+  if (file_name == "downloads/redatam/CP2012BOL/BaseMunicipio_V3/CPV2012Municipio.dic" |
+      file_name == "downloads/redatam/CP2012BOL/BaseMunicipio_V3/CPV2012Comunidad.dicx") {
+    d_sex$sex <- c(2L, 1L)
+  }
 
   # Bind columns and calculate differences for sex
   d_control_sex <- d_control_sex %>%
@@ -184,8 +194,28 @@ generate_tables_r <- function(d_control, country_code, census_year, var_sex, var
 # p08 = sex
 # p09 = age
 
+result <- generate_tables(
+  d_control,
+  "downloads/redatam/CP2017CHL/open-redatam-dic-to-csv/PERSONA.csv.gz",
+  152L,
+  2017L,
+  "p08",
+  "p09"
+)
+
+result2 <- generate_tables_r(
+  d_control,
+  "downloads/redatam/CP2017CHL/BaseOrg16/CPV2017-16.dic",
+  152L,
+  2017L,
+  "p08",
+  "p09"
+)
+
+all.equal(result$d_control_sex, result2$d_control_sex)
+all.equal(result$d_control_age, result2$d_control_age)
+
[…]
