+++ rpkg/R/utils.R
@@ -1,75 +1,57 @@
-detect_encoding_ <- function(strings) {
-  valid_strings <- paste(strings[!is.na(strings) & strings != ""],
-    collapse = " ")
-  if (length(valid_strings) > 0) {
-    detected_encoding <- stri_enc_detect(valid_strings)
-    if (length(detected_encoding) > 0) {
-      return(detected_encoding[[1]]$Encoding[1])
-    }
-  }
-  return("UTF-8")
-}
-
 list_to_datatable_ <- function(x) {
-  for (y in names(x)) {
-    if (is.list(x[[y]])) {
-      x[[y]] <- as.data.table(x[[y]])
-    }
-  }
-  return(x)
+  lapply(x, function(y) {
+    if (is.list(y)) as.data.table(y) else y
+  })
 }
 
-fix_encoding_recursive_ <- function(x, encoding) {
+# Encoding fix ----
+
+fix_encoding_recursive_ <- function(x) {
   if (is.list(x)) {
-    return(lapply(x, fix_encoding_recursive_, encoding))
+    return(lapply(x, fix_encoding_recursive_))
   } else if (is.character(x)) {
-    y <- iconv(x, from = encoding, to = "UTF-8", sub = "")
-    y <- gsub("\u00c3\u00a1", "\u00e1", y) # a with acute accent
-    y <- gsub("\u00c3\u00a9", "\u00e9", y) # e with acute accent
-    y <- gsub("\u00c3\u00b3", "\u00f3", y) # o with acute accent
-    y <- gsub("\u00c3\u00ba", "\u00fa", y) # u with acute accent
-    y <- gsub("\u00c3\u00b1", "\u00f1", y) # n with tilde
-    y <- gsub("\u00b1", "\u00f1", y) # n with tilde
-    y <- gsub("\u00c3", "\u00ed", y) # i with acute accent
-    return(y)
+    replacements <- c(
+      "\u00c3\u00a1" = "\u00e1", # a with acute accent
+      "\u00c3\u00a9" = "\u00e9", # e with acute accent
+      "\u00c3\u00b3" = "\u00f3", # o with acute accent
+      "\u00c3\u00ba" = "\u00fa", # u with acute accent
+      "\u00c3\u00b1" = "\u00f1", # n with tilde
+      "\u00b1" = "\u00f1", # n with tilde
+      "\u00c3" = "\u00ed" # i with acute accent
+    )
+    return(str_replace_all(stri_enc_toutf8(x), replacements))
+  }
+  x
+}
+
+fix_encoding_in_data_table <- function(y) {
+  char_cols <- names(y)[vapply(y, is.character, logical(1))]
+  y[, (char_cols) := lapply(.SD, fix_encoding_recursive_), .SDcols = char_cols]
+  y
+}
+
+# Extract the unique values of the "chr" vectors in the lists
+extract_strings <- function(y) {
+  if (is.data.table(y)) {
+    return(unlist(y[, lapply(.SD, function(col) if (is.character(col)) col else NULL)]))
   } else {
-    return(x)
+    return(unlist(y[vapply(y, is.character, logical(1))]))
   }
 }
 
 fix_encoding_ <- function(x) {
-  # Extract the unique values of the "chr" vectors in the lists
-  strings <- unique(unlist(lapply(x, function(y) {
-    if (is.data.table(y)) {
-      return(unlist(y[, lapply(
-        .SD,
-        function(col) {
-          if (is.character(col)) col else NULL
-        }
-      )]))
-    } else {
-      return(y[vapply(y, is.character, logical(1))])
-    }
-  })))
+  strings <- unique(unlist(lapply(x, extract_strings)))
 
-  # Detect the encoding of the strings
[…]
