+++ rpkg/R/redatam-package.R
@@ -2,5 +2,7 @@
 #' @keywords internal
 #' @importFrom data.table as.data.table is.data.table setnames `:=` `.SD`
 #' @importFrom janitor make_clean_names
-#' @importFrom stringi stri_enc_detect
+#' @importFrom stringi stri_enc_toutf8
+#' @importFrom stringr str_trim str_replace_all
+#' @importFrom tibble as_tibble
 "_PACKAGE"
