### 2024-11-15 6f5fcce
sending to CRAN with fix suggested by Ivan Krylov
