### 2024-11-18 3973ff1
Implement Gabor's suggestion to check R build (https://github.com/r-hub/containers/issues/78#issuecomment-2472862776)
