+++ .github/workflows/build-rpkg.yml
@@ -57,48 +57,8 @@ jobs:
       - name: Checkout
         uses: actions/checkout@v4
 
-      - name: Setup pandoc
-        if: matrix.config.container == null
-        uses: r-lib/actions/setup-pandoc@v2
-
-      - name: Setup R
-        if: matrix.config.container == null
-        uses: r-lib/actions/setup-r@v2
+      - name: Run R CMD check
+        uses: r-hub/actions/run-check@v2
         with:
-          r-version: ${{ matrix.config.r }}
-          http-user-agent: ${{ matrix.config.http-user-agent }}
-          use-public-rspm: true
-
-      - name: Restore R package cache
-        uses: actions/cache@v4
-        with:
-          path: ~/.cache/R
-          key: ${{ runner.os }}-r-${{ matrix.config.r }}-cache
-
-      - name: Install libpng (macOS)
-        if: matrix.config.os == 'macos-14' && matrix.config.r == 'devel'
-        run: |
-          curl -LO https://mac.r-project.org/bin/darwin20/arm64/libpng-1.6.38-darwin.20-arm64.tar.xz
-          sudo tar -xvf libpng-1.6.38-darwin.20-arm64.tar.xz -C /
-          echo "PKG_CONFIG_PATH=/opt/R/arm64/lib/pkgconfig" >> $GITHUB_ENV
-
-      - name: Install dependencies
-        run: |
-          Rscript -e 'if (!requireNamespace("pak", quietly = TRUE)) { install.packages("pak", repos = "https://r-lib.github.io/p/pak/dev/") }'
-          Rscript -e 'pak::pkg_install("decor", dependencies = TRUE)'
-          Rscript -e 'pak::pkg_install("rcmdcheck", dependencies = TRUE)'
-          Rscript -e 'pak::pkg_install("pkgbuild", dependencies = TRUE)'
-          Rscript -e 'pak::pkg_install(".", dependencies = TRUE)'
-        working-directory: ./rpkg
-        
-      - name: Check
-        run: |
-          Rscript -e 'pkgbuild::check_build_tools(debug = TRUE)'
-          Rscript -e 'rcmdcheck::rcmdcheck(args = c("--no-manual", "--as-cran"), build_args = "--no-manual", error_on = "error")'
-        working-directory: ./rpkg
-
-      - name: Save R package cache
-        uses: actions/cache@v4
-        with:
-          path: ~/.cache/R
-          key: ${{ runner.os }}-r-${{ matrix.config.r }}-cache
+          working-directory: ./rpkg
+          
