### 2024-11-18 458343f
install rcmdcheck
