+++ .github/workflows/build-rpkg.yml
@@ -57,7 +57,9 @@ jobs:
       - name: Checkout
         uses: actions/checkout@v4
 
+      - name: Install dependencies
+        run: |
+          R -q -e 'pak::pkg_install(c("deps::.", "any::rcmdcheck"), dependencies = TRUE)'
+
       - name: Run R CMD check
-        uses: r-hub/actions/run-check@v1
-        with:
-          working-directory: ./rpkg
+        uses: r-lib/actions/check-r-package@v2
