+++ .github/workflows/build-rpkg.yml
@@ -13,7 +13,6 @@ permissions: read-all
 jobs:
   R-CMD-check:
     runs-on: ${{ matrix.config.os }}
-    container: ${{ matrix.config.container }}
 
     name: ${{ matrix.config.name }}
 
@@ -21,34 +20,15 @@ jobs:
       fail-fast: false
       matrix:
         config:
-          - {os: macos-13, r: 'release', container: null, name: 'macos-13-r-release'}
-          - {os: macos-14, r: 'release', container: null, name: 'macos-14-r-release'}
-          - {os: ubuntu-latest, r: 'release', container: null, name: 'ubuntu-latest-r-release'}
-          - {os: windows-latest, r: 'release', container: null, name: 'windows-latest-r-release'}
-          - {os: macos-13, r: 'devel', container: null, name: 'macos-13-r-devel'}
-          - {os: macos-14, r: 'devel', container: null, name: 'macos-14-r-devel'}
-          - {os: ubuntu-latest, r: 'devel', container: null, name: 'ubuntu-latest-r-devel'}
-          - {os: windows-latest, r: 'devel', container: null, name: 'windows-latest-r-devel'}
-          - {os: ubuntu-latest, container: { image: 'ghcr.io/r-hub/containers/ubuntu-gcc12:latest' }, name: 'r-devel-linux-x86_64-debian-gcc'}
-          - {os: ubuntu-latest, container: { image: 'ghcr.io/r-hub/containers/ubuntu-clang:latest' }, name: 'r-devel-linux-x86_64-debian-clang'}
-          - {os: ubuntu-latest, container: { image: 'ghcr.io/r-hub/containers/atlas:latest' }, name: 'ATLAS'}
-          - {os: ubuntu-latest, container: { image: 'ghcr.io/r-hub/containers/c23:latest' }, name: 'C23'}
-          - {os: ubuntu-latest, container: { image: 'ghcr.io/r-hub/containers/clang-asan:latest' }, name: 'clang-ASAN'}
-          - {os: ubuntu-latest, container: { image: 'ghcr.io/r-hub/containers/clang16:latest' }, name: 'clang16'}
-          - {os: ubuntu-latest, container: { image: 'ghcr.io/r-hub/containers/clang17:latest' }, name: 'clang17'}
-          - {os: ubuntu-latest, container: { image: 'ghcr.io/r-hub/containers/clang18:latest' }, name: 'clang18'}
-          - {os: ubuntu-latest, container: { image: 'ghcr.io/r-hub/containers/clang19:latest' }, name: 'clang19'}
-          - {os: ubuntu-latest, container: { image: 'ghcr.io/r-hub/containers/clang20:latest' }, name: 'clang20'}
-          - {os: ubuntu-latest, container: { image: 'ghcr.io/r-hub/containers/donttest:latest' }, name: 'donttest'}
-          - {os: ubuntu-latest, container: { image: 'ghcr.io/r-hub/containers/gcc13:latest' }, name: 'gcc13'}
-          - {os: ubuntu-latest, container: { image: 'ghcr.io/r-hub/containers/gcc13:latest' }, name: 'gcc14'}
-          - {os: ubuntu-latest, container: { image: 'ghcr.io/r-hub/containers/intel:latest' }, name: 'intel'}
-          - {os: ubuntu-latest, container: { image: 'ghcr.io/r-hub/containers/mkl:latest' }, name: 'mkl'}
-          - {os: ubuntu-latest, container: { image: 'ghcr.io/r-hub/containers/nold:latest' }, name: 'noLD'}
[…]
