### 2024-11-18 97b64c1
change dir for docker images
