### 2024-11-18 47ef6a3
refactor r readers
