+++ rpkg/src/redatamlib/readers/FuzzyEntityParser.cpp
@@ -1,12 +1,15 @@
-#include <algorithm> //  std::replace
-
-#include <cpp11/function.hpp> //  cpp11::stop
+#include <cpp11/function.hpp>  // stop
 
 #include "FuzzyEntityParser.hpp"
 #include "FuzzyVariableParser.hpp"
-#include "utils.hpp" //  GetFileExtension, ThrowIfBad
+#include "utils/utils.hpp"  //  GetFileExtension, ThrowIfBad
 
 namespace RedatamLib {
+
+  using std::move;
+  using std::make_pair;
+  using cpp11::stop;
+
 FuzzyEntityParser::FuzzyEntityParser(const string &filePath)
     : m_reader(filePath), m_rootPath(FindRootPath(filePath)) {}
 
@@ -19,7 +22,7 @@ vector<Entity> FuzzyEntityParser::ParseEntities() {
     while (true) {
       curr = TryGetEntity();
       if (curr.first) {
-        ret.push_back(curr.second);
+        ret.push_back(move(curr.second));
       } else {
         m_reader.MovePos(1);
       }
@@ -28,7 +31,7 @@ vector<Entity> FuzzyEntityParser::ParseEntities() {
   }
 
   if (ret.empty()) {
-    cpp11::stop("Error: No entities found.");
+    stop("Error: No entities found.");
   }
 
   for (Entity &e : ret) {
@@ -44,45 +47,43 @@ vector<Entity> FuzzyEntityParser::ParseEntities() {
 }
 
 std::pair<bool, Entity> FuzzyEntityParser::TryGetEntity() {
-  size_t ogPos = m_reader.GetPos();
+  size_t originalPosition = m_reader.GetPos();
 
-  string entityName("");
+  string entityName;
   if (!m_reader.TryReadStr(&entityName) || entityName.empty()) {
-    m_reader.SetPos(ogPos);
-    return std::pair<bool, Entity>(false, Entity());
+    m_reader.SetPos(originalPosition);
+    return {false, Entity()};
   }
 
-  string entityNameRepeated("");
+  string entityNameRepeated;
   if (!m_reader.TryReadStr(&entityNameRepeated)) {
     m_reader.MovePos(2);
   } else if (entityName != entityNameRepeated) {
-    m_reader.SetPos(ogPos);
-    return std::pair<bool, Entity>(false, Entity());
+    m_reader.SetPos(originalPosition);
+    return {false, Entity()};
   }
 
-  string parentEntityName("");
+  string parentEntityName;
   if (!entityNameRepeated.empty() && !m_reader.TryReadStr(&parentEntityName)) {
   }
 
-  string description("");
+  string description;
   if (!m_reader.TryReadStr(&description, false)) {
     m_reader.MovePos(2);
   }
 
-  string idxFileName(""), ext("");
+  string idxFileName, ext;
   if ((!m_reader.TryReadStr(&idxFileName, false)) ||
       !(!idxFileName.empty() && TryGetFileExtension(idxFileName, &ext) &&
         ".ptr" == ext)) {
-    m_reader.SetPos(ogPos);
-    return std::pair<bool, Entity>(false, Entity());
+    m_reader.SetPos(originalPosition);
+    return {false, Entity()};
   }
   idxFileName = ReplaceRootPath(m_rootPath, idxFileName);
 
-  std::pair<size_t, size_t> bounds(ogPos, m_reader.GetPos());
+  std::pair<size_t, size_t> bounds = make_pair(originalPosition, m_reader.GetPos());
 
-  return std::pair<bool, Entity>(
-      true,
-      Entity(entityName, parentEntityName, description, idxFileName, bounds));
[…]
