+++ rpkg/src/redatamlib/readers/XMLParser.hpp
@@ -2,9 +2,9 @@
 #define REDATAMLIB_XMLPARSER_HPP
 
 #include <iostream>
-#include <memory> //  std::shared_ptr
+#include <memory>  // shared_ptr
 #include <string>
-#include <utility> //  std::pair
+#include <utility> // pair
 #include <vector>
 
 #include "Entity.hpp"
@@ -18,7 +18,7 @@ using std::string;
 using std::vector;
 
 class XMLParser {
-public:
+ public:
   XMLParser() = default;
   ~XMLParser() = default;
 
@@ -27,7 +27,7 @@ public:
 
   vector<Entity> ParseFile(const string &fileName);
 
-private:
+ private:
   string m_rootPath;
   string GetTagValue(pugi::xml_node node, const string &tag, size_t idx = 0);
   pugi::xml_node ParseEntity(vector<Entity> *results, pugi::xml_node node,
@@ -37,6 +37,6 @@ private:
   string ParseVarRange(pugi::xml_node var);
   vector<Tag> ParseVarTags(pugi::xml_node var);
 };
-} // namespace RedatamLib
+}  // namespace RedatamLib
 
-#endif //  REDATAMLIB_XMLPARSER_HPP
+#endif  //  REDATAMLIB_XMLPARSER_HPP

[…]
