### 2024-11-18 c0c4aed
refactor r entitities
