+++ rpkg/src/redatamlib/entities/Variable.cpp
@@ -5,6 +5,9 @@
 
 namespace RedatamLib {
 
+using std::move;
+using std::make_shared;
+
 Variable::Variable()
     : m_name(""),
       m_type(NA),
@@ -25,7 +28,7 @@ Variable::Variable(const string &name, VarType type, const string &idxFileName,
       m_dataSize(dataSize),
       m_filter(filter),
       m_range(range),
-      m_tags(tags),
+      m_tags(move(tags)),
       m_description(description),
       m_decimals(decimals) {
   ParseValues();
@@ -49,7 +52,7 @@ string Variable::GetDescription() const { return m_description; }
 
 size_t Variable::GetDecimals() const { return m_decimals; }
 
-std::shared_ptr<void> Variable::GetValues() const { return m_values; }
+shared_ptr<void> Variable::GetValues() const { return m_values; }
 
 void Variable::ParseValues() {
   ByteArrayReader reader(m_idxFileName);
@@ -85,13 +88,8 @@ void Variable::ParseValues() {
   }
 }
 
-void Variable::ParseStrings(size_t length, ByteArrayReader reader) {
-  vector<string> *vals = new vector<string>();
-  struct Deleter {
-    void operator()(void *ptr) const {
-      delete static_cast<vector<string> *>(ptr);
-    }
-  };
+void Variable::ParseStrings(size_t length, ByteArrayReader &reader) {
+  shared_ptr<vector<string>> vals = make_shared<vector<string>>();
 
   try {
     while (true) {
@@ -100,16 +98,11 @@ void Variable::ParseStrings(size_t length, ByteArrayReader reader) {
   } catch (const std::out_of_range &) {
   }
 
-  m_values = shared_ptr<void>(static_cast<void *>(vals), Deleter());
+  m_values = vals;
 }
 
-void Variable::ParseIntegers(size_t size, ByteArrayReader reader) {
-  vector<uint32_t> *vals = new vector<uint32_t>();
-  struct Deleter {
-    void operator()(void *ptr) const {
-      delete static_cast<vector<uint32_t> *>(ptr);
-    }
-  };
+void Variable::ParseIntegers(size_t size, ByteArrayReader &reader) {
+  shared_ptr<vector<uint32_t>> vals = make_shared<vector<uint32_t>>();
 
   try {
     switch (size) {
@@ -128,16 +121,11 @@ void Variable::ParseIntegers(size_t size, ByteArrayReader reader) {
   } catch (const std::out_of_range &) {
   }
 
-  m_values = shared_ptr<void>(static_cast<void *>(vals), Deleter());
+  m_values = vals;
 }
 
-void Variable::ParseFloats(ByteArrayReader reader) {
-  vector<double> *vals = new vector<double>();
-  struct Deleter {
-    void operator()(void *ptr) const {
-      delete static_cast<vector<double> *>(ptr);
-    }
-  };
+void Variable::ParseFloats(ByteArrayReader &reader) {
+  shared_ptr<vector<double>> vals = make_shared<vector<double>>();
   string temp;
 
   try {
@@ -149,49 +137,39 @@ void Variable::ParseFloats(ByteArrayReader reader) {
   } catch (const std::out_of_range &) {
   }
 
-  m_values = shared_ptr<void>(static_cast<void *>(vals), Deleter());
+  m_values = vals;
 }
 
-void Variable::ParsePCK(size_t size, ByteArrayReader reader) {
-  vector<uint32_t> *vals = new vector<uint32_t>();
-  struct Deleter {
-    void operator()(void *ptr) const {
-      delete static_cast<vector<uint32_t> *>(ptr);
-    }
-  };
[…]
