+++ rpkg/src/redatamlib/exporters/RListExporter.cpp
@@ -7,10 +7,14 @@
 #include "ParentIDCalculator.hpp"
 
 namespace RedatamLib {
-using std::endl;
-using std::ostringstream;
 
-// just to mimic the original CSVExporter
+using std::endl;
+using std::string;
+using std::ostringstream;
+using std::transform;
+using std::vector;
+using namespace cpp11;
+
 ListExporter::ListExporter(const std::string &outputDirectory)
     : m_path(outputDirectory) {
   if (!m_path.empty() && '/' != m_path.back()) {
@@ -18,33 +22,31 @@ ListExporter::ListExporter(const std::string &outputDirectory)
   }
 }
 
-cpp11::list ListExporter::ExportAllR(
-    const std::vector<Entity> &entities) const {
+list ListExporter::ExportAllR(const std::vector<Entity> &entities) const {
   size_t numEntities = entities.size();
-  cpp11::writable::list result(numEntities);
-  cpp11::writable::strings resultNames(numEntities);
+  writable::list result(numEntities);
+  writable::strings resultNames(numEntities);
 
   for (size_t entityIndex = 0; entityIndex < numEntities; ++entityIndex) {
     const Entity &entity = entities[entityIndex];
-    std::string entityName = entity.GetName();
-    std::transform(entityName.begin(), entityName.end(), entityName.begin(),
-                   ::tolower);
+    string entityName = entity.GetName();
+    transform(entityName.begin(), entityName.end(), entityName.begin(), ::tolower);
 
-    std::string exportingEntityMsg = "Exporting " + entityName + "...";
-    cpp11::message(exportingEntityMsg.c_str());
+    string exportingEntityMsg = "Exporting " + entityName + "...";
+    message(exportingEntityMsg.c_str());
 
     size_t numVariables = entity.GetVariables()->size();
-    cpp11::writable::list entityList(numVariables + 2);  // +2 for REF_ID and PARENT_REF_ID
-    cpp11::writable::strings variableNames(numVariables + 2);
+    writable::list entityList(numVariables + 2);  // +2 for REF_ID and PARENT_REF_ID
+    writable::strings variableNames(numVariables + 2);
 
     // Add REF_ID and PARENT_REF_ID columns
     size_t numRows = entity.GetRowsCount();
-    cpp11::writable::integers ref_id_vec(numRows);
-    cpp11::writable::integers parent_ref_id_vec(numRows);
+    writable::integers ref_id_vec(numRows);
+    writable::integers parent_ref_id_vec(numRows);
     ParentIDCalculator pIDCalc(const_cast<Entity *>(&entity));
 
-    std::string ref_id_name = entity.GetName() + "_REF_ID";
-    std::string parent_ref_id_name = entity.GetParentName() + "_REF_ID";
+    string ref_id_name = entity.GetName() + "_REF_ID";
+    string parent_ref_id_name = entity.GetParentName() + "_REF_ID";
 
     for (size_t row = 0; row < numRows; ++row) {
       ref_id_vec[row] = row + 1;
@@ -70,9 +72,8 @@ cpp11::list ListExporter::ExportAllR(
           case PCK:
           case INT:
           case LNG: {
-            auto values =
-                static_cast<std::vector<uint32_t> *>(v.GetValues().get());
-            cpp11::writable::integers rvalues(numRows);
+            vector<uint32_t> *values = static_cast<std::vector<uint32_t> *>(v.GetValues().get());
+            writable::integers rvalues(numRows);
             for (size_t i = 0; i < numRows; i++) {
               rvalues[i] = values->at(i);
             }
@@ -80,22 +81,20 @@ cpp11::list ListExporter::ExportAllR(
             break;
           }
           case CHR: {
-            auto values =
[…]
