### 2024-11-19 e9e8654
refactor r utils
