+++ rpkg/src/redatamlib/utils/utils.hpp
@@ -1,37 +1,40 @@
 #ifndef REDATAMLIB_UTILS_HPP
 #define REDATAMLIB_UTILS_HPP
 
-#include <cstring>   //  std::strerror
-#include <stdexcept> //  std::invalid_argument
+#include <cstring>  // strerror
 #include <string>
-#include <system_error> //  std::error_code, std::generic_category
+#include <system_error>  // error_code, generic_category
 
 namespace RedatamLib {
+
+using std::error_code;
+using std::generic_category;
+using std::strerror;
 using std::string;
 
 string FindRootPath(const string &fileName);
 string ReplaceRootPath(const string &rootPath, const string &fileName);
 
-//  Throws std::invalid_argument
 string GetFileExtension(const string &fileName);
 
 bool TryGetFileExtension(const string &fileName, string *output);
 
-template <typename E> void ThrowIfBad(bool is_good_, std::error_code err_,
-  const string &e_msg_) {
+template <typename E>
+void ThrowIfBad(bool is_good_, std::error_code err_, const string &e_msg_) {
   if (!is_good_) {
     throw E(err_, e_msg_);
   }
 }
 
-template <typename E> void ThrowIfBad(bool is_good_, int errno_) {
+template <typename E>
+void ThrowIfBad(bool is_good_, int errno_) {
   if (!is_good_) {
-    throw E(std::error_code(errno_, std::generic_category()),
-            std::strerror(errno_));
+    throw E(error_code(errno_, generic_category()), strerror(errno_));
   }
 }
 
-template <typename E> void ThrowIfBad(bool is_good_, const E &e_) {
+template <typename E>
+void ThrowIfBad(bool is_good_, const E &e_) {
   if (!is_good_) {
     throw e_;
   }
@@ -44,6 +47,6 @@ void ThrowIfBad(bool is_good_, const std::string &e_msg_) {
   }
 }
 
-} // namespace RedatamLib
+}  // namespace RedatamLib
 
-#endif //  REDATAMLIB_UTILS_HPP
+#endif  // REDATAMLIB_UTILS_HPP
