+++ rpkg/src/main.cpp
@@ -1,18 +1,17 @@
 #include <string>
 
-#include <cpp11/function.hpp> // cpp11::stop
+#include <cpp11/function.hpp>
 
-#include "redatamlib/entities/RedatamDatabase.hpp"
+#include "redatamlib/database/RedatamDatabase.hpp"
 
 using namespace cpp11;
-using std::string;
 
-[[cpp11::register]] cpp11::list
-export_redatam_to_list_(std::string dicFilePath) {
+[[cpp11::register]] cpp11::list export_redatam_to_list_(std::string dicFilePath) {
   try {
     RedatamLib::RedatamDatabase db(dicFilePath);
     return db.ExportRLists();
   } catch (const std::exception &e) {
-    cpp11::stop("Export failed: %s", e.what());
+    std::string errorMsg = "Error: " + std::string(e.what());
+    cpp11::stop(errorMsg.c_str());
   }
 }
