+++ rpkg/src/Makevars.in
@@ -12,10 +12,10 @@ SOURCES = vendor/pugixml.cpp \
           redatamlib/readers/FuzzyVariableParser.cpp \
           redatamlib/exporters/ParentIDCalculator.cpp \
           redatamlib/exporters/RListExporter.cpp \
-          redatamlib/entities/RedatamDatabase.cpp \
           redatamlib/entities/Variable.cpp \
           redatamlib/readers/XMLParser.cpp \
-          redatamlib/utils.cpp \
+          redatamlib/utils/utils.cpp \
+          redatamlib/database/RedatamDatabase.cpp \
           main.cpp \
           cpp11.cpp
 
