+++ rpkg/DESCRIPTION
@@ -1,7 +1,7 @@
 Package: redatam
 Type: Package
 Title: Import 'REDATAM' Files 
-Version: 2.0.4
+Version: 2.1.0
 Authors@R: c(
     person(
         given = "Mauricio",
