### 2024-11-19 cc3199c
refactor standalone app
