+++ src/readers/ByteArrayReader.cpp
@@ -1,75 +1,87 @@
-#include <algorithm> //  std::transform, std::all_of, std::search
-#include <cctype>    //  std::isalnum
-#include <fstream>   //  std::ifstream, std::ios::binary
-#include <ios>       //  std::ios_base::failure
-#include <iterator>  //  std::ostream_iterator, std::istreambuf_iterator
-#include <sstream>   //  std::ostringstream
+#include <algorithm>  // search
+#include <fstream>    // ifstream
+#include <iterator>   // ostream_iterator, istreambuf_iterator
+#include <sstream>    // ostringstream
 
 #include "ByteArrayReader.hpp"
-#include "utils.hpp"
+#include "utils/utils.hpp"
 
 namespace RedatamLib {
+
+using std::all_of;
+using std::bad_alloc;
+using std::copy;
+using std::exception;
+using std::ifstream;
+using std::isalnum;
+using std::istreambuf_iterator;
+using std::ostream_iterator;
+using std::ostringstream;
+using std::out_of_range;
+using std::search;
+using std::string;
+using std::transform;
+
 ByteArrayReader::ByteArrayReader() : m_currPos(0), m_endPos(0) {}
 
 ByteArrayReader::ByteArrayReader(const string &filePath)
     : m_data(), m_currPos(0), m_endPos(0) {
   string eMsg = "Error: Failed to open file: ";
-  std::ifstream fs(filePath, std::ios::binary);
+  ifstream fs(filePath, std::ios::binary);
   if (!fs.is_open()) {
     // Extract the filename part
 #ifdef _WIN32
-    std::string filename = filePath.substr(filePath.find_last_of("\\") + 1);
-    std::string directory = filePath.substr(0, filePath.find_last_of("\\") + 1);
+    string filename = filePath.substr(filePath.find_last_of("\\") + 1);
+    string directory = filePath.substr(0, filePath.find_last_of("\\") + 1);
 #else
-    std::string filename = filePath.substr(filePath.find_last_of("/") + 1);
-    std::string directory = filePath.substr(0, filePath.find_last_of("/") + 1);
+    string filename = filePath.substr(filePath.find_last_of("/") + 1);
+    string directory = filePath.substr(0, filePath.find_last_of("/") + 1);
 #endif
 
     // Try with lowercase extension, keeping directory intact
-    std::string lowerFilename = filename;
+    string lowerFilename = filename;
     auto pos = lowerFilename.find_last_of('.');
-    if (pos != std::string::npos) {
-      std::transform(lowerFilename.begin() + pos, lowerFilename.end(),
-                     lowerFilename.begin() + pos, ::tolower);
+    if (pos != string::npos) {
+      transform(lowerFilename.begin() + pos, lowerFilename.end(),
+                lowerFilename.begin() + pos, ::tolower);
     }
     fs.open(directory + lowerFilename, std::ios::binary);
     if (!fs.is_open()) {
       // Try with uppercase extension
-      std::string upperFilename = filename;
-      if (pos != std::string::npos) {
-        std::transform(upperFilename.begin() + pos, upperFilename.end(),
-                       upperFilename.begin() + pos, ::toupper);
+      string upperFilename = filename;
+      if (pos != string::npos) {
+        transform(upperFilename.begin() + pos, upperFilename.end(),
+                  upperFilename.begin() + pos, ::toupper);
       }
       fs.open(directory + upperFilename, std::ios::binary);
       if (!fs.is_open()) {
         // Try with uppercase filename and lowercase extension
-        std::string upperFileNameLowerExt = filename;
-        if (pos != std::string::npos) {
-          std::transform(upperFileNameLowerExt.begin(),
[…]
