+++ include/readers/FuzzyVariableParser.hpp
@@ -3,7 +3,7 @@
 
 #include <mutex>
 #include <string>
-#include <utility> //  std::pair
+#include <utility>  //  std::pair
 #include <vector>
 
 #include "ByteArrayReader.hpp"
@@ -16,7 +16,7 @@ using std::string;
 using std::vector;
 
 class FuzzyVariableParser {
-public:
+ public:
   //  throws std::ios_base::failure if fails to open file
   FuzzyVariableParser(const string &filePath);
 
@@ -28,12 +28,12 @@ public:
 
   void ParseAllVariables(vector<Entity> &entities);
 
-private:
+ private:
   ByteArrayReader m_reader;
   string m_rootPath;
   std::mutex m_mtx;
 
-  vector<pair<size_t, size_t>> GetSearchBounds(vector<Entity> entities);
+  vector<pair<size_t, size_t>> GetSearchBounds(vector<Entity> &entities);
 
   static VarType ParseType(ByteArrayReader *reader);
   static string ParseIdxFileName(const string &rootPath,
@@ -43,7 +43,7 @@ private:
   static void ParseMissingAndNA(vector<Tag> *tags, ByteArrayReader *reader);
   static size_t ParseDecimals(ByteArrayReader *reader);
   static size_t GetSubstringLength(string delimiter, ByteArrayReader *reader);
-  static void ParseVariables(shared_ptr<vector<Variable>> output,
+  static void ParseVariables(std::shared_ptr<vector<Variable>> output,
                              pair<size_t, size_t> bounds,
                              const string &rootPath, ByteArrayReader reader);
   static void ThreadParseVars(std::mutex &mutex, size_t start, size_t end,
@@ -52,6 +52,6 @@ private:
                               const string &rootPath, ByteArrayReader reader);
 };
 
-} // namespace RedatamLib
+}  // namespace RedatamLib
 
-#endif //  REDATAMLIB_FUZZYVARIABLEPARSER_HPP
+#endif  //  REDATAMLIB_FUZZYVARIABLEPARSER_HPP
