+++ src/exporters/ParentIDCalculator.cpp
@@ -1,6 +1,7 @@
 #include "ParentIDCalculator.hpp"
 
 namespace RedatamLib {
+
 ParentIDCalculator::ParentIDCalculator(Entity *child)
     : m_child(child), m_currID(0), m_currLimit(0) {}
 
@@ -13,4 +14,4 @@ size_t ParentIDCalculator::GetParentID(size_t currRow) {
   return m_currID;
 }
 
-} // namespace RedatamLib
+}  // namespace RedatamLib

[…]
