### 2024-11-19 48e14ed
tidy std:: in standalone app
