+++ src/utils/utils.cpp
@@ -11,6 +11,9 @@
 
 namespace RedatamLib {
 
+using std::invalid_argument;
+using std::string;
+
 string FindRootPath(const string &fileName) {
 #ifdef _WIN32
   size_t pos = fileName.find_last_of('\\');
@@ -56,7 +59,7 @@ string ReplaceRootPath(const string &rootPath, const string &fileName) {
 string GetFileExtension(const string &fileName) {
   size_t pos = fileName.find_last_of('.');
   if (pos == string::npos) {
-    ThrowIfBad<std::invalid_argument>(false, "Error: File has no extension.");
+    ThrowIfBad<invalid_argument>(false, "Error: File has no extension.");
   }
   return fileName.substr(pos);
 }
@@ -71,21 +74,21 @@ bool TryGetFileExtension(const string &fileName, string *output) {
 }
 
 #ifdef _WIN32
-bool Exists(const std::string &path) {
+bool Exists(const string &path) {
   DWORD attr = GetFileAttributes(path.c_str());
   return (attr != INVALID_FILE_ATTRIBUTES);
 }
 
-bool CreateDirectories(const std::string &path) {
+bool CreateDirectories(const string &path) {
   return CreateDirectory(path.c_str(), NULL) ||
          GetLastError() == ERROR_ALREADY_EXISTS;
 }
 #else
-bool Exists(const std::string &path) {
+bool Exists(const string &path) {
   return access(path.c_str(), F_OK) != -1;
 }
 
-bool CreateDirectories(const std::string &path) {
+bool CreateDirectories(const string &path) {
   mode_t mode = 0755;
   return mkdir(path.c_str(), mode) == 0 || errno == EEXIST;
 }
