+++ include/entities/Variable.hpp
@@ -1,9 +1,9 @@
 #ifndef REDATAMLIB_VARIABLE_HPP
 #define REDATAMLIB_VARIABLE_HPP
 
-#include <memory>  // std::shared_ptr
+#include <memory>  // shared_ptr
 #include <string>
-#include <utility>  // std::pair
+#include <utility>  // pair
 #include <vector>
 
 namespace RedatamLib {
