### 2024-11-19 e97fda7
cmd and tidy code
