+++ rpkg/src/main.cpp
@@ -6,7 +6,8 @@
 
 using namespace cpp11;
 
-[[cpp11::register]] cpp11::list export_redatam_to_list_(std::string dicFilePath) {
+[[cpp11::register]] cpp11::list
+export_redatam_to_list_(std::string dicFilePath) {
   try {
     RedatamLib::RedatamDatabase db(dicFilePath);
     return db.ExportRLists();
