### 2024-11-19 fb4be72
update paths to build on windows
