+++ pypkg/redatamlib/readers/FuzzyEntityParser.hpp
@@ -1,25 +1,25 @@
-//  replaces DataBlock.cs
 #ifndef REDATAMLIB_FUZZYENTITYPARSER_HPP
 #define REDATAMLIB_FUZZYENTITYPARSER_HPP
-
-#include <string>
-#include <unordered_map>
-#include <utility> //  std::pair
-#include <vector>
 
 #include "ByteArrayReader.hpp"
 #include "Entity.hpp"
 
+#include <string>
+#include <unordered_map>
+#include <utility> // pair
+#include <vector>
+
 namespace RedatamLib {
+
 using std::pair;
 using std::string;
 using std::unordered_map;
 using std::vector;
 
 class FuzzyEntityParser {
-public:
+ public:
   //  throws std::ios_base::failure if fails to open file
-  FuzzyEntityParser(const string &filePath);
+  explicit FuzzyEntityParser(const string &filePath);
   ~FuzzyEntityParser() = default;
 
   FuzzyEntityParser(const FuzzyEntityParser &) = delete;
@@ -27,7 +27,7 @@ public:
 
   vector<Entity> ParseEntities();
 
-private:
+ private:
   ByteArrayReader m_reader;
   string m_rootPath;
 
@@ -39,6 +39,6 @@ private:
                              unordered_map<string, Entity *> mapping);
 };
 
-} // namespace RedatamLib
+}  // namespace RedatamLib
 
-#endif //  REDATAMLIB_FUZZYENTITYPARSER_HPP
+#endif  //  REDATAMLIB_FUZZYENTITYPARSER_HPP
