### 2024-11-20 0371290
fix clang-specific note about std::move() vs using std::move + move()
