+++ rpkg/src/redatamlib/readers/XMLParser.cpp
@@ -76,7 +76,7 @@ xml_node XMLParser::ParseEntity(vector<Entity> *results, xml_node node,
   shared_ptr<vector<Variable>> variables = ParseVariables(node);
   curr.AttachVariables(variables);
 
-  results->push_back(move(curr));
+  results->push_back(std::move(curr));
 
   xml_node child = node.child("entity");
   return child;
