+++ pypkg/redatamlib/entity/Entity.cpp
@@ -1,9 +1,7 @@
 #include "Entity.hpp"
 
 namespace RedatamLib {
-
 using std::make_shared;
-using std::move;
 
 Entity::Entity()
     : m_name(""), m_parentName(""), m_description(""), m_indexFilename(""),
@@ -50,7 +48,6 @@ Entity *Entity::GetChild() const { return m_child; }
 void Entity::AttachChild(Entity *child) { m_child = child; }
 
 void Entity::AttachVariables(shared_ptr<vector<Variable>> variables) {
-  m_variables = move(variables);
+  m_variables = std::move(variables);
 }
-
 } // namespace RedatamLib
