+++ rpkg/src/redatamlib/readers/FuzzyEntityParser.cpp
@@ -4,10 +4,8 @@
 #include <cpp11/function.hpp> // stop
 
 namespace RedatamLib {
-
 using cpp11::stop;
 using std::make_pair;
-using std::move;
 using std::out_of_range;
 using std::pair;
 
@@ -23,7 +21,7 @@ vector<Entity> FuzzyEntityParser::ParseEntities() {
     while (true) {
       curr = TryGetEntity();
       if (curr.first) {
-        ret.push_back(move(curr.second));
+        ret.push_back(std::move(curr.second));
       } else {
         m_reader.MovePos(1);
       }
@@ -98,5 +96,4 @@ void FuzzyEntityParser::AssignChildren(
     }
   }
 }
-
 } // namespace RedatamLib
