+++ src/entities/Variable.cpp
@@ -4,7 +4,6 @@
 
 namespace RedatamLib {
 using std::make_shared;
-using std::move;
 using std::out_of_range;
 
 Variable::Variable()
@@ -16,7 +15,8 @@ Variable::Variable(const string &name, VarType type, const string &idxFileName,
                    vector<Tag> tags, const string &description, size_t decimals)
     : m_name(name), m_type(type), m_idxFileName(idxFileName),
       m_dataSize(dataSize), m_filter(filter), m_range(range),
-      m_tags(move(tags)), m_description(description), m_decimals(decimals) {
+      m_tags(std::move(tags)), m_description(description),
+      m_decimals(decimals) {
   ParseValues();
 }
 
