### 2024-11-26 7466b58
re-fix relative path + tidy readme
