### 2024-11-27 #672 @emilyriederer
Hi @pachadotdev - thanks for your patience as we discussed internally.

This looks like a great project that is undoubtedly extremely useful and well within the rOpenSci scope. However, we feel overall we need to adhere to the rOpenSci policy of not reviewing forks due to the complexities that could cause in downstream transparency, maintainability, etc. For example, changes is one project are less likely to be fully tested for the other; one part of the project could change it's license; etc.

I think either solutions where the full codebase is contained in one repo or another solution which limits yours to the current R package and uses the C++ project as a dependency could work.

If you might be interested in restructuring, I'd happily hold this and plan for a full review of the new project. Otherwise, it may be best to close for now.
