### 2024-11-28 #672 @emilyriederer
Thanks @pachadotdev for the replies. Just to clarify: I understand that you and @litalbaraki have established a good working model that sounds very effective for your purpose. I understand the complexities here and the concern isn't so much the operating model of _this specific package_ but general concerns that illustrate why we have the current policy.
