### 2024-12-03 #672 @emilyriederer
Hi @pachadotdev ! Thanks for the update.

We do review submissions on branches, but that case assumes that packages post-review will ultimately be hosted on the default GitHub branch as the sole and primary instance. If both you and @litalbaraki are happy with that, this might work. Otherwise, I can run this back by the team to discuss
