### 2024-12-03 #672 @pachadotdev
> Hi @pachadotdev ! Thanks for the update.
>
> We do review submissions on branches, but that case assumes that packages post-review will ultimately be hosted on the default GitHub branch as the sole and primary instance. If both you and @litalbaraki are happy with that, this might work. Otherwise, I can run this back by the team to discuss

@litalbarkai are you ok sending your repo to submission? I was considering merging my fork into your repo, but all the tutorials were McGyver level and I gave up.
