### 2024-12-03 #672 @pachadotdev
> @pachadotdev @litalbarkai [This StackOverflow chunk](https://stackoverflow.com/a/10548919) can be used to merge two repos. I just tried it for your repos, and everything worked perfectly. You can also easily transfer issues across, if you need. Because the two repos are in different orgs, you might need to create a dummy repo in origin-org, transfer all issues to that, transfer dummy repo to destination-org, then move issues from there to destination-repo.

can I keep my fork just in case?

using https://github.com/litalbarkai/open-redatam/tree/main/rpkg seems to be ok (that is the original repo)
