### 2024-12-10 #672 @emilyriederer
Per discussion [here](https://github.com/ropensci/software-review/issues/645#issuecomment-2532191271), we're moving this to "hold" state while `Capybara` proceeds
