### 2025-02-02 #672 @maurolepore
@pachadotdev, FYI my EiC rotation just started. I see this issue is on hold. If the conditions are such that you're able to resume just let me know or whoever is the EiC at the time.

Just as a reminder of our policy:

> The holding status will be revisited every 3 months, and after one year the issue will be closed. -- https://devguide.ropensci.org/softwarereview_policies.html#policiesreviewprocess
