+++ rpkg/DESCRIPTION
@@ -1,7 +1,7 @@
 Package: redatam
 Type: Package
 Title: Import 'REDATAM' Files 
-Version: 2.1.0
+Version: 2.1.1
 Authors@R: c(
     person(
         given = "Mauricio",
@@ -38,9 +38,9 @@ Suggests:
     rmarkdown,
     testthat (>= 3.0.0)
 Depends: R(>= 3.5.0)
-Description: Import 'REDATAM' formats into R via the 'Open REDATAM' C++ library
-    <https://github.com/litalbarkai/open-redatam> based on De Grande (2016)
-    <https://www.jstor.org/stable/24890658>.
+Description: Import 'REDATAM' formats into R via the 'Open REDATAM' C++ library.
+    The full context of this project and details about the implementation are
+    available in <doi:10.1017/dap.2025.4> (Open Access).
 License: Apache License (>= 2)
 URL: https://github.com/litalbarkai/open-redatam
 BugReports: https://github.com/litalbarkai/open-redatam/issues
