+++ rpkg/DESCRIPTION
@@ -1,7 +1,7 @@
 Package: redatam
 Type: Package
 Title: Import 'REDATAM' Files 
-Version: 2.1.1
+Version: 2.1.2
 Authors@R: c(
     person(
         given = "Mauricio",
