+++ .github/workflows/build-windows.yml
@@ -16,6 +16,10 @@ jobs:
       - name: Checkout code
         uses: actions/checkout@v4
 
+      - name: Set environment variables
+        run: |
+          echo "AQTSERVER=https://ftp.jaist.ac.jp/pub/qtproject/" >> $GITHUB_ENV
+
       - name: Install Qt
         uses: jurplel/install-qt-action@v4
         with:
