+++ .github/workflows/build-windows.yml
@@ -19,7 +19,7 @@ jobs:
       - name: Install Qt
         uses: jurplel/install-qt-action@v4
         with:
-          version: '5.15.*'
+          version: '5.15.16'
           host: 'windows'
           target: 'desktop'
           arch: 'win64_msvc2019_64'
