### 2025-04-17 f43e36b
try qt 5.15.2
