### 2025-05-04 #672 @maurolepore
@pachadotdev this is to mark the end of my EiC roation and to hand it over to the next EiC with a small summary.

* This is on hold since the author has another open submission ([here](https://github.com/ropensci/software-review/issues/645))

Thanks!
