### 2025-06-19 486e446
switch to uv, update docs
