+++ pypkg/redatam.egg-info/requires.txt
@@ -1,3 +1,0 @@
-pandas
-numpy
-chardet
