### 2025-06-19 ca3180f
update gh workflow for pypkg
