### 2025-06-19 f676628
update workflow to follow astral docs
