+++ .github/workflows/build-windows.yml
@@ -10,7 +10,7 @@ on:
 
 jobs:
   build-windows:
-    runs-on: windows-2019
+    runs-on: windows-2025
 
     steps:
       - name: Checkout code
