### 2025-07-22 4ce300a
Update test-windows-installer.yml
