+++ .github/workflows/test-windows-installer.yml
@@ -10,7 +10,7 @@ on:
 
 jobs:
   test-installer:
-    runs-on: windows-latest
+    runs-on: windows-2025
 
     steps:
       - name: Checkout code
