### 2025-07-22 1b1038c
Update test-windows-installer.yml
