+++ .github/workflows/test-windows-installer.yml
@@ -10,7 +10,7 @@ on:
 
 jobs:
   test-installer:
-    runs-on: windows-2025
+    runs-on: windows-latest
 
     steps:
       - name: Checkout code
