+++ .github/workflows/build-windows.yml
@@ -10,7 +10,7 @@ on:
 
 jobs:
   build-windows:
-    runs-on: windows-2025
+    runs-on: windows-latest
 
     steps:
       - name: Checkout code
