### 2025-07-22 93dd895
Update cmake visual studio generator version
