+++ .github/workflows/build-windows.yml
@@ -34,10 +34,12 @@ jobs:
       - name: Install NSIS
         run: choco install nsis -y
 
+      - name: Configure CMake
+        run: cmake -G "Visual Studio 17 2022" .
+
       - name: Build redatam
         run: |
           cd redatamwindows
-          cmake -G "Visual Studio 16 2019" .
           cmake --build . --config Release
           cmake --install . --config Release
           cd ..
@@ -45,7 +47,6 @@ jobs:
       - name: Build redatamgui
         run: |
           cd redatamguiwindows
-          cmake . -G "Visual Studio 16 2019"
           cmake --build . --config Release
           & "C:\Qt\5.15.2\msvc2019_64\bin\windeployqt.exe" --release .\Release\redatamgui.exe
           cp ..\redatamwindows\Release\redatam.exe Release\redatam.exe
