### 2025-07-22 376a7fa
Update build-windows.yml
windows action runner is set to latest
cmake generator is set to VS2022
