+++ .github/workflows/build-windows.yml
@@ -10,7 +10,7 @@ on:
 
 jobs:
   build-windows:
-    runs-on: windows-2019
+    runs-on: windows-latest
 
     steps:
       - name: Checkout code
@@ -37,7 +37,7 @@ jobs:
       - name: Build redatam
         run: |
           cd redatamwindows
-          cmake -G "Visual Studio 16 2019" .
+          cmake -G "Visual Studio 17 2022" .
           cmake --build . --config Release
           cmake --install . --config Release
           cd ..
@@ -45,7 +45,7 @@ jobs:
       - name: Build redatamgui
         run: |
           cd redatamguiwindows
-          cmake . -G "Visual Studio 16 2019"
+          cmake -G "Visual Studio 17 2022" .
           cmake --build . --config Release
           & "C:\Qt\5.15.2\msvc2019_64\bin\windeployqt.exe" --release .\Release\redatamgui.exe
           cp ..\redatamwindows\Release\redatam.exe Release\redatam.exe
