### 2025-08-11 87d2413
parsing old (old old) DIC files for Chile 1992
