### 2026-02-06 #672 @pachadotdev
> Hi [@pachadotdev](https://github.com/pachadotdev) , as discussed [here](https://github.com/ropensci/software-review/issues/645#issuecomment-3834506714), we will now pause Capybara which allows us to consider this again. In the intervening time, have you had any more thoughts on how to manage the unique branching structure?
>
> To recap, when we review packages we assume that packages post-review will ultimately be hosted on the default GitHub branch as the sole and primary instance.

hi!

I am not sure. Should I move the rpkg branch to its own repo?
