### 2026-02-07 #672 @emilyriederer
I know there are a lot of complexities you are thinking through and need alignment with your co-author, but I think if you find that to be a reasonable solution for your repo maintenance/development, that would certainly create a clean and clear path from our side.
