### 2026-02-08 #672 @emilyriederer
Thank you! As you characterize it as a "mirror" repo, just to clarify: will both be maintained going forward? Which will be the "official" respository, wrt links on CRAN, etc.?
