### 2026-02-10 #672 @pachadotdev
> Thanks! Maybe a better question: which would be the primary development repository?

can we agree on this?

library https://github.com/pachadotdev/open-redatam
R pkg https://github.com/pachadotdev/open-redatam-rpkg
