### 2026-02-11 #672 @emilyriederer
Thanks @pachadotdev ! Could you please explain a bit more on the relationship between the two repos? Is the first the C++ code and the second the R package which wraps the C++ code?
