### 2026-02-20 #672 @mpadge
@pachadotdev Just to be explicitly clear: As you've seen [elsewhere](https://github.com/ropensci-review-tools/roreviewapi/issues/64), our checking systems can now handle packages in sub-directories. In response to that, could you please:

1. Edit the YAML metadata in the initial template above to update `"Repository:"` to just the direct repository URL.
2. Ensure that that repository remains the one and only true source for the package during the process.

Once you've updated the template, please call `@ropensci-review-bot check package`, and we'll proceed from there. Thanks!
