+++ rpkg/src/redatamlib/readers/FuzzyEntityParser.cpp
@@ -2,6 +2,9 @@
 #include "FuzzyVariableParser.hpp"
 #include "utils.hpp"          //  GetFileExtension, ThrowIfBad
 #include <cpp11/function.hpp> // stop
+#include <cctype> // isdigit
+#include <algorithm> // std::min
+#include <regex>
 
 namespace RedatamLib {
 using cpp11::stop;
@@ -13,6 +16,12 @@ FuzzyEntityParser::FuzzyEntityParser(const string &filePath)
     : m_reader(filePath), m_rootPath(FindRootPath(filePath)) {}
 
 vector<Entity> FuzzyEntityParser::ParseEntities() {
+  // Check if this is an old format dictionary file
+  if (IsOldFormat()) {
+    return ParseEntitiesOldFormat();
+  }
+
+  // Use the original parsing logic for newer formats
   pair<bool, Entity> curr;
   vector<Entity> ret;
   unordered_map<string, Entity *> entities;
@@ -94,6 +103,272 @@ void FuzzyEntityParser::AssignChildren(
     if (!parent.empty()) {
       mapping[parent]->AttachChild(&e);
     }
+  }
+}
+
+bool FuzzyEntityParser::IsOldFormat() {
+  // Save current position
+  size_t originalPos = m_reader.GetPos();
+
+  try {
+    // Reset to beginning of file
+    m_reader.SetPos(0);
+
+    // Read a sample of the file (first 2000 characters)
+    size_t sampleSize = std::min((size_t)2000, m_reader.GetEndPos());
+    string sample = m_reader.ReadString(sampleSize);
+
+    // Restore original position
+    m_reader.SetPos(originalPos);
+
+    // Old format characteristics:
+    // 1. Long lines with multiple embedded references
+    // 2. Contains multiple PTR file references
+    // 3. Contains DATASET keywords
+    // 4. Entities and data are on same line
+    size_t ptrCount = 0;
+    size_t pos = 0;
+
+    while ((pos = sample.find(".PTR", pos)) != string::npos) {
+      ptrCount++;
+      pos += 4;
+    }
+    bool hasDataset = sample.find("DATASET") != string::npos;
+    bool hasMultiplePtrs = ptrCount > 1;
+    bool longFirstLine = sample.find('\n') == string::npos || sample.find('\n') > 1000;
+    return hasDataset && hasMultiplePtrs && longFirstLine;
+  } catch (const std::exception &) {
+    // Restore position and assume new format
+    m_reader.SetPos(originalPos);
+    return false;
+  }
+}
+
+vector<Entity> FuzzyEntityParser::ParseEntitiesOldFormat() {
+  vector<Entity> ret;
+  size_t originalPos = m_reader.GetPos();
+  m_reader.SetPos(0);
+
+  try {
+    // Read the entire file content safely
+    size_t fileSize = m_reader.GetEndPos();
+    string content = m_reader.ReadString(fileSize);
+
+    // Define the expected entities and their PTR file patterns
+    vector<std::pair<string, string>> entityPatterns = {
+      {"CL92", "0000.PTR"},
+      {"REGION", "0002.PTR"}, 
+      {"PROVINC", "0005.PTR"},
+      {"COMUNA", "0008.PTR"},
+      {"DISTRIT", "0011.PTR"},
+      {"AREA", "0014.PTR"},
+      {"ZONALOC", "0017.PTR"},
+      {"MANZENT", "0020.PTR"},
+      {"SECTOR", "0024.PTR"},
+      {"VIVIEND", "0026.PTR"},
+      {"HOGAR", "0047.PTR"},
+      {"PERSONA", "0073.PTR"}
+    };
+
[…]
