### 2026-02-20 e5b104d
v2.2.0 for rOpenSci review
