+++ rpkg/R/cpp4r.R
@@ -1,0 +1,5 @@
+# Generated by cpp4r: do not edit by hand
+
+export_redatam_to_list_ <- function(dicFilePath) {
+	.Call(`_redatam_export_redatam_to_list_`, dicFilePath)
+}
