+++ rpkg/R/cpp11.R
@@ -1,5 +1,0 @@
-# Generated by cpp11: do not edit by hand
-
-export_redatam_to_list_ <- function(dicFilePath) {
-  .Call(`_redatam_export_redatam_to_list_`, dicFilePath)
-}
