+++ rpkg/src/cpp4r.cpp
@@ -1,16 +1,16 @@
-// Generated by cpp11: do not edit by hand
+// Generated by cpp4r: do not edit by hand
 // clang-format off
 
 
-#include "cpp11/declarations.hpp"
+#include "cpp4r/declarations.hpp"
 #include <R_ext/Visibility.h>
 
 // main.cpp
-cpp11::list export_redatam_to_list_(std::string dicFilePath);
+cpp4r::list export_redatam_to_list_(std::string dicFilePath);
 extern "C" SEXP _redatam_export_redatam_to_list_(SEXP dicFilePath) {
-  BEGIN_CPP11
-    return cpp11::as_sexp(export_redatam_to_list_(cpp11::as_cpp<cpp11::decay_t<std::string>>(dicFilePath)));
-  END_CPP11
+  BEGIN_CPP4R
+    return cpp4r::as_sexp(export_redatam_to_list_(cpp4r::as_cpp<cpp4r::decay_t<std::string>>(dicFilePath)));
+  END_CPP4R
 }
 
 extern "C" {

[…]
