### 2026-02-28 #672 @emilyriederer
Hi @pachadotdev !

Thanks for the explanation. Helpful to understand the complexities of the file format. I think additionally, when @mpadge spoke about "explanation", in addition to motivation he was also referring to shedding more light on the code structure. (As well as having a website, as he mentions).

I started to dig into the code structure piece myself as part of this conversation. I was struck with a better understanding of the interplay between your R and C code. I note that in the `R/` directory:

- There's ~230 LOC (R) total in 2 files
- The main [read_redatam.R](https://github.com/litalbarkai/open-redatam/blob/main/rpkg/R/read_redatam.R) (60 LOC) doesn't really have much to review independently as it's just a thin wrapper that defines the external function, in turn calling the non-exported functions one-by-one
- So, the only real R code is in [utils.R](https://github.com/litalbarkai/open-redatam/blob/main/rpkg/R/utils.R) (170 LOC) which is a lot of cleaning/reformatting of nested objects (I assume a lot of the heavy lifting of what you describe above!) which are likely more binary "it works or it doesn't" unit-testable functions than big picture design choices

This made me wonder if you have any thoughts on what parts of the package you think would benefit most from review or if there are specific areas you're hoping a review would focus on?
