### 2026-03-04 #672 @emilyriederer
Hi @pachadotdev !

Thanks so much for the continued engagement and context sharing on this issue.

Unfortunately, after discussing internally, we do not feel equipped to take up a review of `redatam` at this time. The package seems to demand a lot of reviewers at the moment since, to @mpadge 's point, it doesn't provide reviewers with clarifying documentation that provides rationale for some of the design decisions. At the same time, our reviewers often drive a lot of value by helping authors raise the bar on R code related design decisions, and the scope of the `utils.R` helpers feels a bit too small and tactical for that review.

I will close out this issue, but I want to thank you as always for your submissions and continued involvement in the rOpenSci commity!
